expand scope of version check (#893)
* expand scope of version check

expand scope of version checking script to look at the other files with "version:" fields. namely, all of the macro files, and docs.md

* rename script

rename not_module_has_version.sh to be more descriptive of what it actually does (checks if file is either docs or one of the macros).