change subdirectory naming convention
suggest we use `code` instead of `src` for the subdirectory to contain scripts, notebooks, etc as this is a more self-explanatory name for learners who are new to this field if they're poking around in the repo trying to find things. (Not a single one of our modules currently follows this rule anyways, so it will be the same amount of work to bring into compliance either way).